pyproject.toml
[project]
name = "genai-accelerator-ai-assistant"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "nylas==6.9.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
